% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading.R
\name{download.example.data}
\alias{download.example.data}
\title{Downloads example data files for package demonstration.}
\usage{
download.example.data(dest = ".", allow.cache = T,
  base.url = "http://freyja.cs.cofc.edu/downloads/GeneListPrioritization")
}
\arguments{
\item{dest}{The destination path}

\item{allow.cache}{Allow the files to be cached to prevent downloading multiple times.}

\item{base.url}{Source URL location}
}
\description{
Downloads example data files for package demonstration.
}
\examples{
download.example.data()
}

