% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.genes.R
\name{get.all.genes.kegg}
\alias{get.all.genes.kegg}
\title{Downloading gene ids from all Kegg Pathways}
\usage{
get.all.genes.kegg(sleep = 10, species = "hsa", start = 1,
  file = "kegg.genes.csv")
}
\arguments{
\item{sleep}{Number of seconds to sleep between Kegg requests}

\item{species}{Species identifier for Kegg. Default is hsa.}

\item{start}{Index of pathway to start processing. Only necessary because of frequent Kegg timeouts.}

\item{file}{File to read gene ids from and}
}
\description{
Writes the gene ids to a file
}
\examples{
gene.ids = get.all.genes.kegg(start=1)
}

