% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.genes.R
\name{get.all.genes.wiki}
\alias{get.all.genes.wiki}
\title{Downloading all genes in all WikiPathways pathways}
\usage{
get.all.genes.wiki(url = "http://webservice.wikipathways.org",
  species = "Homo sapiens")
}
\arguments{
\item{url}{WikiPathways base URL for webservice calls.}

\item{species}{Species for pathway analysis}
}
\value{
a dataframe
}
\description{
Returns the gene ids in WikiPathways for a given species
For more information on WikiPathways see http://webservice.wikipathways.org/ and
http://www.wikipathways.org/index.php/Help:WikiPathways_Webservice/API
}
\examples{
gene.ids = get.all.genes.wiki(webg.pathways)
}

