% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.genes.R
\name{get.genes.in.data}
\alias{get.genes.in.data}
\title{Filter for only those genes in the data.}
\usage{
get.genes.in.data(gene.ids, gene.data)
}
\arguments{
\item{gene.ids}{A vector of Entrez Gene IDs}

\item{gene.data}{The gene data loaded with the helper functions.}
}
\value{
a vector of Gene IDs
}
\description{
Returns a corrected gene ID vector
}
\examples{
download.example.data()
Cattaneo.rna = load.gene.data("Cattaneo_array.csv",5)
webg.pathways = load.WebGestalt("Marra_0_kegg_protein_enrichment.tsv",'Kegg')
gene.ids = get.genes.kegg(webg.pathways)
genes.in.data = get.genes.in.data(gene.ids,Cattaneo.rna$data)
}

