% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.genes.R
\name{get.genes.tf}
\alias{get.genes.tf}
\title{Finds the genes ids from the Transcription Factor Targets database}
\usage{
get.genes.tf(webg.pathways, db = system.file("extdata",
  "c3.tft.v4.0.entrez.gmt", package = "ComplementaryDomainPrioritization"))
}
\arguments{
\item{webg.pathways}{dataframe where the second column contains the pathway ids}

\item{db}{TF database to use}
}
\value{
an array of gene ids
}
\description{
Returns ...
}
\examples{
download.example.data()
webg.pathways = load.WebGestalt("Marra_0_tf_protein_enrichment.tsv",'TF')
gene.ids = get.genes.tf(webg.pathways)
}

