% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.genes.R
\name{get.genes.wiki}
\alias{get.genes.wiki}
\title{Downloading WikiPathways and returning gene ids}
\usage{
get.genes.wiki(webg.pathways, url = "http://webservice.wikipathways.org",
  species = "Homo sapiens")
}
\arguments{
\item{webg.pathways}{Dataframe where the second column contains the pathway ids}

\item{url}{WikiPathways base URL for webservice calls.}

\item{species}{Species for pathway analysis}
}
\value{
a dataframe
}
\description{
Returns ...
For more information on WikiPathways see http://webservice.wikipathways.org/ and
http://www.wikipathways.org/index.php/Help:WikiPathways_Webservice/API
}
\examples{
download.example.data()
webg.pathways = load.WebGestalt("Marra_0_wiki_protein_enrichment.tsv",'Wiki')
gene.ids = get.genes.wiki(webg.pathways)
}

