% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{list.filter.inxs}
\alias{list.filter.inxs}
\title{Returns the gene.ixs for filtering}
\usage{
list.filter.inxs(data, gene.ids)
}
\arguments{
\item{data}{- transposed gene data}

\item{gene.ids}{- result of get.genes.*(...)}
}
\value{
an array of inxs
}
\description{
A helper function that filters using a list of gene IDs and returns an
array of indices of the genes that passed the filter.
}
\examples{
download.example.data()
webg.pathways = load.WebGestalt("Marra_0_wiki_protein_enrichment.tsv",'Wiki')
gene.ids = get.genes.wiki(webg.pathways)
Catteno.rna = load.gene.data("Catteno_array.csv",5)
gene.ixs = list.filter.inxs(Catteno.rna$transposed.data,gene.ids)
}

