% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{list.filter}
\alias{list.filter}
\title{Returns a filtered dataset}
\usage{
list.filter(data, gene.ids)
}
\arguments{
\item{data}{- transposed gene data}

\item{gene.ids}{- result of get.genes.*(...)}
}
\value{
an array of inxs
}
\description{
Returns ...
}
\examples{
download.example.data()
webg.pathways = load.WebGestalt("Marra_0_wiki_protein_enrichment.tsv",'Wiki')
gene.ids = get.genes.wiki(webg.pathways)
Catteno.rna = load.gene.data("Catteno_array.csv",5)
filtered.data = list.filter(Catteno.rna$transposed.data,gene.ids)
}

