% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading.R
\name{load.gene.data}
\alias{load.gene.data}
\title{Loading a correctly formatted gene file.}
\usage{
load.gene.data(file, start.data.inx)
}
\arguments{
\item{file}{A string pointing to the location of the data file}

\item{start.data.inx}{An integer >= 2 that indicates where the actual data starts.}
}
\value{
a list containing two data frames.
}
\description{
Returns a list containing two data frames.
The second data frame is a formatted and transposed version
of the original data with an additional labels column added.
For data format instructions see https://github.com/Anderson-Lab/ComplementaryDomainPrioritization/wiki.
}
\examples{
download.example.data()
TCGA.rna = load.gene.data("TCGA_rna.csv",3)
Cattaneo.rna = load.gene.data("Cattaneo_array.csv",5)
}

