% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading.R
\name{load.WebGestalt}
\alias{load.WebGestalt}
\title{Loading the pathway data from WebGestalt output.}
\usage{
load.WebGestalt(file, db)
}
\arguments{
\item{file}{A string pointing to the location of the data file}

\item{db}{Wiki, Kegg, or TF}
}
\value{
a list containing the extracted pathways from WebGestalt output.
}
\description{
Returns an array of pathways.
For data format instructions see https://github.com/Anderson-Lab/ComplementaryDomainPrioritization/wiki.
}
\examples{
download.example.data()
pathways = load.WebGestalt("Marra_0_tf_protein_enrichment.tsv",'TF')
pathways = load.WebGestalt("Marra_0_wiki_protein_enrichment.tsv",'Wiki')
pathways = load.WebGestalt("Marra_0_kegg_protein_enrichment.tsv",'Kegg')
}

