% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{overall.mean.filter}
\alias{overall.mean.filter}
\title{Overall mean filter}
\usage{
overall.mean.filter(data, theta)
}
\arguments{
\item{data}{- transposed gene data}

\item{theta}{- if > 1, theta is the target number of ranked variables;
otherwise, it is the top theta fraction of ranked variables.}
}
\value{
an array of inxs
}
\description{
Applies the overall mean filter on the columns of the dataset.
}
\examples{
download.example.data()
webg.pathways = load.WebGestalt("Marra_0_wiki_protein_enrichment.tsv",'Wiki')
gene.ids = get.genes.wiki(webg.pathways)
Cattaneo.rna = load.gene.data("Cattaneo_array.csv",5)
prioritized.data = list.filter(Cattaneo.rna$transposed.data,gene.ids)
prioritized.filtered.data = overall.mean.filter(prioritized.data,100)
}

