% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filters.R
\name{overall.var.filter}
\alias{overall.var.filter}
\title{Overall variance filter}
\usage{
overall.var.filter(data, theta)
}
\arguments{
\item{data}{- transposed gene data}

\item{theta}{- if > 1, theta is the target number of ranked variables;
otherwise, it is the top theta fraction of ranked variables.}
}
\value{
an array of inxs
}
\description{
Performs an overall variance filter using each column.
}
\examples{
download.example.data()
webg.pathways = load.WebGestalt("Marra_0_wiki_protein_enrichment.tsv",'Wiki')
gene.ids = get.genes.wiki(webg.pathways)
Catteno.rna = load.gene.data("Catteno_array.csv",5)
prioritized.data = list.filter(Catteno.rna$transposed.data,gene.ids)
prioritized.filtered.data = overall.var.filter(prioritized.data,100)
}

